+++ server/src/app/views/clients/index.blade.php
@@ -1,24 +1,19 @@
-
 @extends('layouts.master')
 
-@section('sidebar')
-    @parent
-
-    <p></p>
-@stop
-
 @section('content')
+
+@include('partials.theModal')
 
 <div class="row">
   <div class="col-md-6">
     <div class="box box-success">
       <div class="box-body no-padding">
-        <table class="table table-condensed">
+        <table class="table table-condensed table-hover">
           <thead>
             <tr>
               <th style="width: 10px">#</th>
               <th>Nombre</th>
-              <th style="width: 80px">Acciones</th>
+              <th class="text-right">Acciones</th>
             </tr>
           </thead>
           <tbody>
@@ -26,9 +21,9 @@
             <tr>
                 <td>{{ $index + 1}}.</td>
                 <td><a href="{{ URL::route('clientss.classrooms.index', $client->id) }}">{{$client->name}}</a></td>
-                <td>
-                  <a href="#" class="label label-danger destroy" data-client-id="{{$client->id}}"><i class="fa fa-trash-o"></i></a>
-                  <a href="{{ URL::route('clientss.edit', $client->id) }}" class="label label-success"><i class="fa fa-pencil"></i></a>
+                <td class="text-right">
+                  <a href="#" class="label label-danger destroy" data-client-id="{{$client->id}}" title="Seguro?"><i class="fa fa-trash-o"></i></a>
+                  <a href="#" class="label label-success" data-toggle="modal" data-target="#theModal" data-remote="{{ URL::route('clientss.edit', $client->id) }}"><i class="fa fa-pencil"></i></a>
                 </td>
             </tr>
             @endforeach
@@ -45,21 +40,33 @@
 
 @section("js")
 @parent
+{{ HTML::script('js/vendor/bootstrap-confirmation.js') }}
+
 <script>
- $(document).ready(function() {
-      $(".destroy").on("click",function(){
-        var $this = $(this);
-        var clientId = $this.data("client-id");
-        var APIurl = '{{ URL::route('clientss.index') }}' + '/' + clientId;
-        $.ajax({
-          url: APIurl,
-          type: "DELETE",
-          success: function(res) {
-            console.log(res);
-            $this.closest('tr').remove();
-          }
-       });
-     });
- });
+$(document).ready(function() {
+  $('.destroy').confirmation({
+    btnCancelLabel: 'Cancelar',
+    btnOkLabel: 'Eliminar',
+    onConfirm: function(event, element){
+      var clientId = element.data("client-id");
+      var APIurl = '{{ URL::route('clientss.index') }}' + '/' + clientId;
+      $.ajax({
+        url: APIurl,
+        type: "DELETE",
+        success: function(res) {
+          console.log(res);
+          element.closest('tr').remove();
+        }
+      });
+    }
+  });
+
+  $('body').on('hidden.bs.modal', '.modal', function () {
+    var $this = $(this);
+    $this.removeData('bs.modal');
+    $this.find('.modal-title').html('Cargando...');
+    $this.find('.modal-body').html('<h1 style="text-align: center;"><i class="fa fa-spinner fa-spin"></i></h1>');
+  });
+});
 </script>
 @stop
